# frontend/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ApiProblem" : {
        "properties" : {
          "detail" : {
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "timestamp" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        },
        "required" : [ "detail", "status", "timestamp", "title", "type" ],
        "type" : "object"
      },
...
      "AuthResponse" : {
        "properties" : {
          "accessToken" : {
            "type" : "string"
          },
          "expiresInSeconds" : {
            "format" : "int64",
            "type" : "integer"
          },
          "tokenType" : {
            "type" : "string"
          },
          "user" : {
            "$ref" : "#/components/schemas/UserResponse"
          }
        },
        "required" : [ "accessToken", "expiresInSeconds", "tokenType", "user" ],
        "type" : "object"
      },
...
      "LoginRequest" : {
        "properties" : {
          "email" : {
            "format" : "email",
            "minLength" : 1,
            "type" : "string"
          },
          "password" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "email", "password" ],
        "type" : "object"
      },
      "MfaRequiredResponse" : {
        "properties" : {
          "message" : {
            "type" : "string"
          },
          "mfaToken" : {
            "type" : "string"
          }
        },
        "required" : [ "message", "mfaToken" ],
        "type" : "object"
      },
      "MfaVerifyRequest" : {
        "properties" : {
          "code" : {
            "minLength" : 1,
            "type" : "string"
          },
          "mfaToken" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "code", "mfaToken" ],
        "type" : "object"
      },
      "MonthSummary" : {
        "properties" : {
          "inflow" : {
            "type" : "string"
          },
          "month" : {
            "type" : "string"
          },
          "outflow" : {
            "type" : "string"
          }
        },
        "required" : [ "inflow", "month", "outflow" ],
        "type" : "object"
      },
...
      "RegisterRequest" : {
        "properties" : {
          "email" : {
            "format" : "email",
            "minLength" : 1,
            "type" : "string"
          },
          "fullName" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "password" : {
            "maxLength" : 2147483647,
            "minLength" : 8,
            "type" : "string"
          }
        },
        "required" : [ "email", "fullName", "password" ],
        "type" : "object"
      },
...
      "TotpDisableRequest" : {
        "properties" : {
          "code" : {
            "minLength" : 1,
            "type" : "string"
          },
          "password" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "code", "password" ],
        "type" : "object"
      },
      "TotpEnableRequest" : {
        "properties" : {
          "code" : {
            "minLength" : 1,
            "type" : "string"
          },
          "currentCode" : {
            "type" : "string"
          },
          "currentPassword" : {
            "type" : "string"
          }
        },
        "required" : [ "code" ],
        "type" : "object"
      },
...
      "TransactionResponse" : {
        "properties" : {
          "amount" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "flagged" : {
            "type" : "boolean"
          },
          "fromIban" : {
            "description" : "null on deposits/credits (no originator)",
            "type" : [ "string", "null" ]
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "kind" : {
            "enum" : [ "TRANSFER", "DEPOSIT", "INTEREST" ],
            "type" : "string"
          },
          "memo" : {
            "type" : [ "string", "null" ]
          },
          "postedAt" : {
            "description" : "null while HELD or when CANCELLED - money has not moved",
            "type" : [ "string", "null" ]
          },
          "reviewed" : {
            "type" : "boolean"
          },
          "status" : {
            "enum" : [ "POSTED", "HELD", "CANCELLED" ],
            "type" : "string"
          },
          "toIban" : {
            "description" : "null on engine loan charges (no destination)",
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "amount", "createdAt", "currency", "flagged", "fromIban", "id", "kind", "memo", "postedAt", "reviewed", "status", "toIban" ],
        "type" : "object"
      },
...
      "UserResponse" : {
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "fullName" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "role" : {
            "enum" : [ "CUSTOMER", "ADMIN" ],
            "type" : "string"
          },
          "totpEnabled" : {
            "type" : "boolean"
          }
        },
        "required" : [ "email", "fullName", "id", "role", "totpEnabled" ],
        "type" : "object"
      }
    },
    "securitySchemes" : {
      "bearer" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Full-stack online banking demo. Errors are RFC-7807 (ApiProblem); money travels as JSON strings.",
    "title" : "Online Banking Simulator API",
    "version" : "1.3.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/accounts/{id}/summary" : {
      "get" : {
        "operationId" : "summary",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "months",
          "required" : false,
          "schema" : {
            "default" : 6,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/MonthSummary"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "transfer-controller" ]
      }
    },
...
    "/api/v1/auth/login" : {
      "post" : {
        "description" : "Returns 200 with an authenticated session, or 202 with a single-use MFA challenge when the account has a factor enabled.",
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            },
            "description" : "Authenticated - session tokens issued"
          },
          "202" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MfaRequiredResponse"
                }
              }
            },
            "description" : "MFA required - verify the returned mfaToken"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Invalid credentials or validation failure"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Wrong email or password"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblem"
                }
              }
            },
            "description" : "Login attempts throttled"
          }
        },
        "summary" : "Log in",
        "tags" : [ "auth-controller" ]
      }
    },
    "/api/v1/auth/logout" : {
      "post" : {
        "operationId" : "logout",
        "parameters" : [ {
          "in" : "cookie",
          "name" : "refresh_token",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "auth-controller" ]
      }
    },
...
    "/api/v1/auth/mfa/verify" : {
      "post" : {
        "operationId" : "mfaVerify",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MfaVerifyRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "auth-controller" ]
      }
    },
    "/api/v1/auth/refresh" : {
      "post" : {
        "operationId" : "refresh",
        "parameters" : [ {
          "in" : "cookie",
          "name" : "refresh_token",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "auth-controller" ]
      }
    },
    "/api/v1/auth/register" : {
      "post" : {
        "operationId" : "register",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "auth-controller" ]
      }
    },
...
    "/api/v1/auth/totp/disable" : {
      "post" : {
        "operationId" : "totpDisable",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TotpDisableRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "auth-controller" ]
      }
    },
    "/api/v1/auth/totp/enable" : {
      "post" : {
        "operationId" : "totpEnable",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TotpEnableRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "auth-controller" ]
      }
    },
...
    "/api/v1/operations" : {
      "get" : {
        "operationId" : "operation",
        "parameters" : [ {
          "in" : "query",
          "name" : "key",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TransactionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "transfer-controller" ]
      }
    },
...
  },
  "security" : [ {
    "bearer" : [ ]
  } ],
  "servers" : [ {
    "url" : "http://localhost:8080"
  } ]
}